initialcommit

+++ src/app/components/view/view.component.ts
@@ -24,10 +24,13 @@ export class ViewComponent implements OnInit {
 
   delete(id)
   {
-    console.log(id);
-    this.authService.deleteuser(id).subscribe((res:any)=>{
+    if(confirm("Are you sure to delete "))
+    {
+      console.log(id);
+      this.authService.deleteuser(id).subscribe((res:any)=>{
       console.log(res);
       window.location.reload();
     })
+    }    
   }
 }
